EDX mid-course project
This is a mid-course project presented for the Python for Data Science course in EDX.
In this project we use an open dataset from Kaggle: The World Development Indicators Dataset. This is a slightly modified version of the original dataset from The World Bank List of the available indicators and a list of the available countries.